Changes to improve calvingMIP radial symmetry
CalvingMIP Experiments 2 and 4 have a prescribed, radially symmetric calving rate.
In the previous code version, retreat was a bit faster along the N/S/E/W axes than along diagonals.
This commit contains several changes that improve radial symmetry and give better agreement
with the prescribed retreat and advance rates.

In the calving module, there is a new subroutine, compute_calving_front_length,
that estimates the CF length in each grid cell. The idea is to draw line segments
between adjacent cells along the CF and compute the length of the segments.
For example, if the CF lies along a row parallel to the x-axis, cf_length = dx,
and if the CF lies along a column parallel to the y-axis, cf_length = dy.
If the CF is parallel to a diagonal (e.g., y = x), then cf_length = sqrt(dx^2 + dy^2).
For some cells, cf_length is the sum of an along-axis segment and a diagonal segment.
There is logic to ensure that each CF cell is connected to at most 2 CF neighbors.

I tried a new method for restricting flow from partial CF cells to unprotected ocean cells.
The old method is based on subroutine redistribute_unprotected_ice; any ice that flows
into unprotected cells is pushed back upstream to the cells where it came from,
using a simplified method to estimate fluxes.
I tried doing this in a more exact way, by passing edgemask_e and edgemask_n to
glissade_horizontal_remap. These are integer masks with a value of 1 for all edges
across which transport is allowed, and otherwise a value of 0.
I set edgemask = 0 for certain edges of partial CF and ocean cells when using
the subgrid CF scheme. I got the method partially working, but in more complex configurations
(e.g., with retreat in CalvingMIP Experiment 4), I couldn't figure out how to avoid
negative areas. The problem is that for some patterns of masking, we can have outflow
across an open edge without having any inflow across the adjacent masked edge.
So I reverted back to the old method.
For now, there is a logical parameter 'use_edgemasks' that can be set to true to activate
the new edgemask method. I may remove this option later, unless I somehow get it fully working.

Minor changes:
* Fixed some initialization logic so that Cp is not set to 0 on restart when running
  with constant Cc or Cp
* Added effective_areafrac to I/O variables
* Replaced some diagnostic prints with calls to subroutine point_diag

With these changes, calvingMIP results look good for Experiments 1 to 4.
It remains to test these changes for physically-based calving laws.