Removed optional edgemasks from incremental remapping
In an earlier commit on this branch, I added optional edge masks to the remapping transport.
The goal was to prevent ice from leaving partial CF cells.
However, the masks did not work, since blocking flow across some edges and not others
leads to negative ice thickness in many simulations. This commit removes the edgemask logic.

I also turned off the verbose_calving option.

This commit is BFB for calvingMIP and other simulations with calving.