Changes to improve CalvingMIP accuracy
Comparing to results from other groups, we found that CISM has too much retreat
during the retreat phase of Experiments 2 and 4, mainly because of the method
used to compute cf_length. I was computing cf_length by reconstructing the CF
along a path connecting cell centers. This path can become noisy (i.e., with
spurious projections and indentations), giving excessive retreat.

I tried two new approaches:
(1) Compute the CF length in a cell based on the number of cell edges bordering the ocean.
    Cells with two ocean edges have their thinning rate multiplied by sqrt(2),
    reflecting the longer CF (but not by a factor of 2, since the calving direction
    is not normal to the two edges).
(2) Compute the CF length based on the secant of the angle between the calving direction
    and the x and y axes. The thinning rate is multiplied by 1 for the compass directions,
    by sqrt(2) for the diagonals, and by intermediate values in between.

Method (1) does not fix the problem of excessive retreat.
Method (2), however, gives more accurate retreat and is now the default for CalvingMIP.

For realistic ice sheet geometries, the calving rate is not radially symmetric,
so method (2) is not appropriate. Method (1) is now the default for subgrid CF schemes
other than CalvingMIP. It is a bit cleaner than the previous method.

Other changes:
(1) Modified the method of estimating the CF location along the diagonal trajectories
    in the circular domain. This method uses the x0 and y0 coordinates and is a bit
    more accurate than the old method; it is less prone to getting stuck for several
    years at the same value.
(2) Added a subroutine to estimate the CF location along the Halbrane and Caprona
    trajectories for the Thule geometry. The Caprona method is approximate and might
    not be as accurate as offline methods using 2D interpolation schemes.
(3) The 1D fields x0 and y0 are now loadable at startup, like x1 and y1.
(4) To deal with I/O involving x0 and y0, I added some code in the distributed get_var
    and put_var functions in parallel_mpi.F90.
(5) Added a config option, which_ho_calvingmip_domain, to specify whether the user
    is running on the circular or Thule domain. This parameter is set automatically
    by the CalvingMip setup script.

Note: This commit will be followed by a commit that includes a modifed CalvingMIP setup script.
      That script creates a grid with the origin at a cell corner instead of a cell center.