Modified CalvingMIP setup script and template
This commit modifies the CalvingMIP setup script.
The main change is that the grid origin is now located at a
cell corner instead of a cell center. This means that on a 5-km grid
measuring 1600 km along each edge, there are now 320 cells in each
direction instead of 321. This change makes it easier to initialize
the calving front at a distance of 750 km from the origin,
instead of being extended or falling short by half a grid cell.

Other changes:
* Fixed a bug that gave the wrong retreat rate for Experiment 4
* Added the which_ho_calvingmip_domain variable to the config files
* Edited the README file for clarity and accuracy