Changed the thck_effective calculation at the calving front
This is the version used for the CISM CalvingMIP submission in Nov. 2024.

This commit changes the way thck_effective (the effective ice thickness)
is computed in calving front cells.

For any CF cell with an interior neighbor, the calculation is unchanged;
thck_effective is set to the maximum thickness of the cell's interior edge neighbors.

For CF cells without interior edge neighbors (i.e., cells whose only edge neighbors
are other CF cells), thck_effective is now set to the maximum thickness of the
CF neighbors. Previously, we set thck_effective = thck, giving values that are too low.

I also fixed a minor bug in the way the CF location is computed along the diagonal axes
for the circular domain.

I set dthck_dx_cf = 0.0005 in the config template file for CalvingMIP.
This value gives slightly more accurate results than the default value of 0.
Setting dthck_dx_cf = 0 gives a little too much retreat.
If desired, we could modify the python setup script to control this parameter
at the command line.

We ran the final CalvingMIP runs with dthck_dx_cf = 0.0005.