Calving changes for Antarctic runs
This commit makes a few changes to improve stress-based calving for Antarctic spin-ups.

(1) The ice speed, which helps determine the lateral calving rate, is now extrapolated
to CF cells as a weighted fraction of the values in the CF cell and the upstream cell.
This weighted extrapolation was already done for the stress eigenvalues.
This change helps reduce oscillations at the CF.

(2) Floating cells with thin ice (H < thck_effective_min), which are adjacent to
other floating cells with thin ice, are now identified as CF cells.
Treating such cell as full cells can lead to numerical issues.

(3) When computing thck_effective based on the thickness of interior neighbors,
cap the neighbor thickness at the flotation thickness. This prevents excessive values
of thck_effective when there are neighboring land-based cells with thick ice.

Note that changes (2) and (3) are answer-changing for CalvingMIP on the Thule domain.

In subroutine glissade_calving_front_mask, I simplified some logic by turning
all the optional arguments into mandatory arguments.

I removed the deprecated cull_calving_front option and associated subroutine.

I also tried out some changes in the damage-based calving. This scheme isn't working as well
as hoped. One reason is that the damage in partial CF cells tends to be smaller than the
damage in upstream cells. This scheme remains under construction.