Added stochastic options for Antarctic calving
This commit introduces two new calving options, both involving random calving events
in which the calving probability is related to the strain rate.

The options for whichcalving (= marine_margin in the config file) now include:
[8]  Deterministic calving based on eigenvalues of the horizontal stress tensor
[9]  Stochastic calving based on eigenvalues of the horizontal stress tensor
[10] Calving based on stochastic internal damage

Option 8, which existed previously, gives good results for idealized (CalvingMIP) and realistic (AIS) tests.
Options 9 and 10 are new. They are technically working but do yet produce realistic AIS CFs.
The previous option 9 (eigencalving) is now option 11 and has not otherwise changed.

Recall that option 8 works as follows:
* In each calving-front (CF) cell, we compute an effective calving stress:
       eff_str = c1*max(tau1,0) + c2*max(tau2,0),
  where typically c2 = 1.0 and c1 < 1.0.
* Based on eff_str, we compute a lateral calving rate L:
       L = (eff_str/str0) * v,
  where str0 is a stress scale and v is the ice speed at the CF.
* We convert C to a thickness change dH in the cell during interval dt:
       dH = -C * dt * H_eff * L_cf / (dx*dy),
  where H_eff is the effective thickness and L_cf is the estimated CF length.

For the AIS, we get good results for str0 ~ 60 kPa and c2 ~ 0.25.
Optionally, we can now introduce a minimum value for L to reduce CF advance in regions
where the modeled eff_str is small and no ice shelves are observed.

Option 9 is similar, but instead of computing a deterministic rate L,
we compute the probability p_c of a calving event of size dx*dy (i.e., one grid cell area)
and thickness H_eff (i.e., the full column) during time interval dt:
       p_c = 1.0 - exp[(eff_str/str0) * v * dt * cf_length / (dx*dy)].
We then generate a random number r between 0 and 1. If r < p_c, there is a calving event.
If H_eff > H (as will often be the case in partial CF cells), then the calving
is extended to the upstream cell(s).

Optionally, we can introduce a minimum eff_str to reduce CF advance in regions
where the modeled eff_str is small and no ice shelves are observed.

We can also control the size of calving events by introducing a parameter calving%length_scale,
which can be thought of as a typical iceberg diameter.
Suppose dx = dy = 8 km and calving_length_scale = 4 km.
We then modify the equations so that there are 4x as many events, each with calving area
equal to 1/4 the grid cell area.

Ideally, option 9 would reduce to option 8 in the limit of a small length scale and many events.
Currently, however, option 9 allows long-term CF advance and retreat where option 8 gives a stable CF.
Need to figure out how to get option 9 to behave more like option 8.

Option 10 is based on a different idea: Stress in the shelf interior induces damage which
is advected toward the CF. Damage is represented as a binary scalar with values of 0 (undamaged)
and 1 (damaged). When a damaged cell reaches the CF, it calves. To vary the size of calving events
in a random way, we also calve any damaged cells that are part of a connected region
with damaged CF cells. This is based on an idea of David Bahr (JGR, 1995).

In practice, the method is tricky to implement. If we compute eff_str in the way above,
we find there is widespread damage in the interior of large shelves (e.g., Filchner),
which spuriously removes much of the shelf.
I tried an alternate method:
     eff_str = max(c1*tau1 + c2*tau2, 0),
which allows large negative values of tau2 to inhibit damage.
We increment the damage probability using
     dD/dt = eff_str/D0, where D0 is an empirical constant with units of Pa*s.
Then we generate random numbers to assign new binary values to D.

A typical pattern consists of:
(1) retreat for some large ice shelves, especially Ross
(2) advance of other large shelves, especially W. Ronne and PIG
(3) spurious advance of small shelves in regions such as Coats, Enderby, and Wilkes Land.
(4) lots of jagged CF edges.
Maybe the results can be improved with further work.

Other changes:
* When removing icebergs while running a subgrid calving scheme, partial CF cells
  are no longer able to spread the fill. Thus, partial CF cells are protected from
  calving only if they border full cells.
* Increased the (somewhat arbitrary) CF length for grid cells with three ocean neighbors;
  now set to (dx + dy) instead of sqrt(dx^2 + dy^2)
* Added a lateral_rate_min parameter for stress-based calving. This allows the user
  to specify a minimum rate in regions where the effective calving stress is small or zero.
* Removed some deprecated calving parameters