Inversion changes, including E inversion with a speed target
This commit makes a number of changes to inversion methods.
The biggest change is a new option to invert for flow_enhancement_factor E
based on a surface speed target, as suggested by Tim van den Akker.
I removed the option to invert for E based on a thickness target.
The options to invert for C_c, C_p and deltaT_ocn based on a thickness target remain.

I fixed a bug in subroutine ismip6_bmlt_float.
This subroutine could change the value of deltaT_ocn even when that field was supposed
to be held fixed. Thanks to Tim for pointing out the bug.

For all thickness-based inversions, I changed the leading (dH/H0) term to (dH/H0)^n,
where n is a user-prescribed exponent.
The default is n = 1, but I have been running with n = 2 to give greater weight
to larger errors dH.

For the which_ho_flotation_function options, I removed the deprecated option 2.
The new option 2 is the old option 3 (using a linear flotation function without
extrapolating value to ice-free ocean cells).
I created a new option 3, which computes the flotation function based on a field
called topg_raised. This is a field in which the coarse-grid bed topography is
higher (i.e., shallower) to better capture pinning points on the BedMachine grid.
We are now testing various ways to compute topg_raised.
I removed an old option to adjust topg on coarse grids based on the standard deviation
of topg as computed on a finer grid.

Other changes:
* When inverting for C_c, we now initialize C_c to the same value everywhere
  (instead of starting with a low value for cells that are floating or ice-free ocean).
* In the inversion for C_c and C_p, I removed two optional terms that weren't being used.
* Changed the default timescales for all inversions to 200 years.
* Replaced some verbose diagnostics with calls to the point_diag subroutine.
* Added some parallel_globalindex diagnostics to the matrix checks in the velocity solver.
* Set verbose_calving = F.