Added Laplacian smoothing terms for inversion
This commit allows the user to add Laplacian smoothing terms when inverting
for basal friction coefficients (Cc or Cp), ocean temperature corrections (deltaT_ocn),
or the flow enhancement factor (E). The goal is to inhibit noisy fields
with large differences between adjacent grid cells.

The Laplacian operator for a scalar H at a cell center is discretized as
    del^2(H) = d^2(H)/dx^2 + d^2(H)/dy^2,
where d^2(H)/dx^2 = (dH/dx|e - dh/dx|w) / dx, where 'e' is the east face,
'w' is the west face, dx is the cell width, and dH/dx is discretized in
the usual way. The y term is analogous.

The new Laplacian terms take the form del^2(H) * L0^2/tau0,
where tau0 is the usual timescale, and L0 is a new length scale,
typically of the same order of magnitude as dx and dy. Larger L0 leads to
more smoothing.

In glissade_grid_operators.F90, I added subroutines for computing the Laplacian
of both cell-centered and staggered variables.

The user can now specify the length scale for each inversion in the config file.
Thee default values are 0, i.e. no smoothing. On an 8km Antarctic grid,
a value L0 = 10 km noticeably smooths the fields with only a small increase
(a few m) in the rmse for H.

I hoped that this change would make the spun-up AIS more robust when inversion
is turned off, but the improvement is modest; more work is still needed.

I also changed the logic in inversion for Cc, Cp, and E.
For these parameters, we expect the physical changes to vary more in proportion
to the logs of these quantities than the absolute values.
The inversion subroutines now convert the Cc (and other quantities) to log10(Cc),
prognose changes in log10(Cc), and convert back to Cc at the end.
This is not very different from what was done before, but the logic is now a bit cleaner.

I set a default timescale of 200 yr for each inversion.

I changed the max and min values of E to 10.0 and 0.1.
At some point, these should be config parameters.

In subroutine glissade_inversion_basal_friction, I set f_ground_weight = F by default.
(Previously, it was true.) With this change, the staggering uses the unweighted thickness
in each cell, instead of weighting by f_ground. This means the inversion might
better discourage thinning of partly grounded cells. In practice, the difference
is small, although thickness biases are reduced in a few local areas.