Added an option to toggle inversion
This commit introduces a new config parameter, inversion_toggle_frequency.
The goal is to make the spin-up more robust by buffering against undesired advance
and retreat when inversion is turned off.

If the toggle frequency = 0. (the default value), then inversion takes place
based on the values of which_ho_coulomb_c, which_ho_deltaT_ocn, etc.
(CISM does inversion when these options are set to 1.)

If the toggle frequency > 0., then the inversion is toggled on and off at regular intervals.
For example, if frequency = 1000 and CISM is configured to do inversion,
then the inversion is turned on during the following time intervals (yr):
[0,1000.], (2000.,3000.], (4000.,5000.], etc.
At other invervals, the inversion is turned off:
(1000.,2000.], (3000.,4000.0], (5000.,6000.], etc.
(Note the parentheses and brackets to indicated open and closed intervals.)

I also introduced a logical variable called inversion_relax_everywhere,
which is true by default but false when toggling the inversion.
* If true, then wherever the usual inversion conditions are not met (e.g., in floating cells
  when we are inverting for C_c), we relax toward the default values.
  The Laplacian smoothing term is also turned on, while the
  terms based on (H - H_obs) and dH/dt are set to 0.
* If false, then the relaxation and Laplacian terms are turned off in cells that do not
  meet the usual inversion conditions, and we retain values computed previously
  when the inversion conditions were satisfied.
  Typically, the previous inverted values will discourage undesired advance and retreat.

I removed the parameter deltaT_ocn_relax, which was always set to zero in previous runs.

I added a parameter deltaT_basin_relax, which is the basin average of deltaT_ocn.
This value is computed at runtime and is uniform within each basin.
For grounded cells in a given basin, deltaT_ocn is now relaxed toward the basin-average value
instead of toward zero.