Removed thck_error_exponent parameter
Several commits ago, I introduced a parameter called thck_error_exponent
with a default value of 1. Setting this parameter to a value of 1
modified the size of the inversion terms based on the thickness bias,
raising the quantity (H - H_obs)/H0 to an arbitrary power.

Tests with values > 1 showed that results did not generally improve;
instead, the overall thickness bias was greater.

With this commit, an exponent of 1 is assumed in all cases.