Changed relaxation and smoothing for inactive cells during inversion
I changed the way inactive cells are treated during inversion.
Here, 'inactive' means that an certain parameter is not physically relevant for a cell.
For example, the value of coulomb_c makes no difference in fully floating cells,
so floating cells are inactive when inverting for coulomb_c.
Similarly, the value of deltaT_ocn makes no difference in fully grounded cells,
so grounded cells are inactive when inverting for deltaT_ocn.

In inactive cells, during inversion, we now zero out terms associated with the thickness error
and dH/dt, but we keep the terms associated with relaxation and Laplacian smoothing.
The values in active cells will then transition smoothly toward the values in inactive cells.
Values in inactive cells will approach the relaxation value when we are far from any active cells.
The goal is to make the spin-up more robust by removing limiting abrupt transitions.

I removed the config option deltaT_ocn_extrapolate. This was used to set deltaT_ocn to the
basin-average value in inactive cells, but is no longer needed if we continually relax
to the basin-average value in these cells.

I also added a config parameter, inversion_phaseout_timescale.
Setting this parameter to a value > 0 increases each of the inversion timescales
(babc_timescale, deltaT_ocn_timescale, flow_enhancement_timescale) by a factor exp(time/phaseout_period).
For instance, if this parameter = 5000 yr, then each timescale increases by a factor of e every 5000 yr,
eventually reaching a large value such that the changes due to inversion are negligible.
The goal is to improve robustness by phasing out the inversion gradually instead of suddenly.