Created a new module, glissade_mass_balance
This commit adds a new module, glissade_mass_balance, in a file of the same name.
This module will take care of calculations related to the surface mass balance
(and some aspects of the basal mass balance).

For now, it contains an initialization subroutine with some operations
previously done in glissade_initialise.

I will add more functionality in upcoming commits, moving code from the
glissade and glissade_transport modules.
This is part of a larger effort to clean up the glissade_module.