Removed the tim0 scale parameter
Following up on the previous commit, this seems a good time to remove
the confusing scale factors from CISM.
This commit removes tim0, which has units of time.
As a result, the timestep dt now has units of seconds instead of being dimensionless.
The old value was equal to the number of seconds in 400 yr.

The method is as follows:
* Comment out every line containing tim0, including the declaration in glimmer_paramets.
* Replace each tim0 operation with the equivalent, assuming tim0 = 1.
* Recompile the code to make sure all instances of tim0 were removed.
* Run some tests to make sure the answers are the same within roundoff.
  In general, answers will not be BFB, since some multiplication and divisions
  will be modified or removed.

This commit fixes a diagnostic mass conservation issue with the previous commit,
which changed the units of bmlt; have to be careful adding bmlt_applied*dt to calving_thck
when bmlt_applied lacks the thk0 scale but calving_thck still includes it.

Later I'll delete the commented-out lines, but I'm leaving them for now
to assist with debugging when we do more extensive testing.

For this commit, the dome test and a glacier test are BFB, while an AIS test
has answer changes within roundoff.