Fixed some thickness roundoff issues
In glissade_inversion_init, there is logic to remove cells with H < thklim
at the start of the run. This is problematic if thklim = 1.0 m and if there
are many cells in the input file with H = 1.0 m. We don't want to remove them,
but they are removed if H is very slightly less than 1.0 m due to rounding.
This commit fixes the issue by adding buffers against roundoff-level differences.

In glissade_get_masks, the logic is that ice_mask = 1 if thck > thklim.
It makes sense to have '>' rather than '>=' in case thklim = 0.
But suppose thklim = 1.0 m, and many cells in the input file also have thck = 1.0 m.
We would like to make these input cells as ice-covered.
This will be the case if we reduce thklim by a very small amount, e.g. 1.0e-11.
We now do this in glissade_initialise:
model%numerics%thklim = max(model%numerics%thklim - eps11, 0.0d0)

In glide_write_diag, there is a threshold parameter, diag_minthck,
which is set to either 0 or thklim. The above adjustment to thklim ensures
that cells with H = 1.0 m (exactly) are included in the diagnostics
when the user sets thklim = 1.0 m in the config file.

These fixes can lead to small answer changes.