Implemented a slope correction for the DIVA solver
Until now, the DIVA velocities have been too large in regions with steep slopes.
This can be understood as follows: The equations are derived with the assumption
that the flow is roughly parallel to the x-axis, which is not the case for steep slopes.
The d/dx(mu du/dx) term in the momentum balance should include a component
perpendicular to the bed, with a proportionality constant of sin^2(theta),
where theta is the angle between the bed and the x-axis.
In the standard DIVA equations, this component is missing, since the x derivative
of the vertically-averaged u is typically small. As a result, DIVA underestimates
the resistive interior stresses and overestimates the ice speed.

A solution is suggested by an analysis of the slab problem presented in Dukowicz (2012, TC).
For this problem, John derived analytic solutions for the velocity based on the Stokes
and BP equations. Studying these solutions, I noticed that the DIVA solution can be
modified to reproduce the analytic BP solution (at least for the slab problem)
by making the following changes:

(1) When assembling the basal BC terms in the stiffness matrix, integrate over the
    actual area of the basal surface (in general, a curved 2D surface embedded in 3D)
    rather than the projection of this surface onto the xy plane.
    This means dividing the projected area by cos(theta), which increases the resistance
    to sliding.
(2) Compute a slope correction term gamma = cos^2(theta) + 4 sin^2(theta).
    Insert this term in the following places:
    (a) Instead of setting beta_eff = beta/(1 + beta*F2),
        set beta_eff = (gamma*beta) / (gamma + beta*F2*cos(theta)), thereby
	increasing beta_eff.
	Recall that beta_eff is defined such that beta_eff * u_mean = beta*u_b,
	so a larger beta_eff implies a smaller u_mean.
    (b) When computing the 3D DIVA velocity profile, multiply the vertical shear
        by cos(theta)/gamma, thus reducing the shear.
    (c) When computing du/dz, set du/dz = (tau_x*sigma)/(efvs*gamma), instead of just
        (tau_x*sigma)/efvs. Here, sigma = (s - z)/H is the unitless vertical coordinate.
    (d) When computing the effective strain rate, multiply the (du/dz)^2 and (dv/dz)^2 terms
        by gamma, thus increasing the effective viscosity for Glen's n > 1.

With these changes, the DIVA speeds match the BP speeds almost exactly for the slab problem.
In general, these corrections work well for slab-like problems where the flow is dominated
by shear perpendicular to the bed. They do not work as well for cases with significant
along-flow tensile or lateral stresses (e.g., ISMIP-HOM test A).

The changes are needed in both the x and y momentum equations. So we compute slope correction terms
for both axes, leading to separate parameters beta_eff_x and beta_eff_y.

To implement the corrections, the user should set diva_slope_correction = .true. in the config file.
The default (at least for now) is .false., which reverts to the standard DIVA solver without slope corrections.
At some point we'll change the default to true.

I used these corrections for the runs in the CISM glacier paper (Minallah, Lipscomb et al., 2025, GMD).
This made it possible to increase the max_slope parameter to 1.0 (i.e., a 45-degree angle)
without getting excessively large flow speeds.

Minor changes and fixes:
* A fix to support periodic BCs in the north-south direction for the slab
* Don't allow Cp to be reset at restart when running with Cp inversion for glaciers
* Option to count the fraction of edges steeper than max_slope