DIVA velocity fix for marine ice
When implementing the DIVA slope correction for glaciers, I changed the DIVA
3D velocity calculation in a way that leads to NaNs for marine-based ice.
This commit fixes the problem. It leads to small answer changes
for land-based ice, compared to the initial DIVA slope correction.
With the correction turned off, the agreement with earlier code
(before the slope correction) is now at roundoff level.