Another DIVA velocity fix
When I changed the DIVA 3D velocity calculation to be correct for large slopes,
I introduced a divzero for floating ice.

Then when I fixed the divzero, I broke the calculation for large slopes.

With this commit, I think the velocity is correct for both floating ice
and large slopes.