Fixed rare remapping error
This commit fixes a rare error that can occur when doing remapping transport
for very thin ice layers. The error happens when, within machine precision,
the total column thickness is greater than zero, but the thickness of individual
layers is not.

The fix is to call make_remap_mask for each layer with thck_layer as an input argument,
instead of calling it just once with thck as an argument.

Results are BFB except in the rare cases when errors used to be triggered.

Thanks to Tim van den Akker for diagnosing the problem and figuring out the fix.