Turned off Laplacian smoothing of coulomb_c for grounded ice
An earlier commit added a Laplacian smoothing term to the equations for inversion
of coulomb_c, powerlaw_c, deltaT_ocn, and flow_enhancement factor.
The goal was to make the inversion more robust by having smoother transitions
in ice properties near grounding lines.

I found that in Antarctic spin-ups, applying the Laplacian term to coulomb_c
can cause oscillations. Instead of converging in 25 to 30 iterations, the nonlinear solver
fails to converge. This significantly slows down the code, from < 1 hr per 1000 model years
to about 1:45 per 1000 model years on 256 Derecho cores.

This commit sets the Laplacian term to zero in the C_c and C_p inversions for cells
that are at least partly grounded. The Laplacian term is applied only in floating cells
and ice-free ocean, where it is balanced by a relaxation term.
Thus, we still have a fairly smooth transition in C_c or C_p when the GL advances downstream,
but downstream floating cells no longer have a smoothing influence on upstream grounded cells.

With this change, performance improves. The time to run 1000 yr is again < 1 hr.

Applying the Laplacian term to inversion of deltaT_ocn and flow_enhancement_factor
does not significantly slow down the code, so I left the Laplacian term in those equations
(at least for now).

I tried a 10 kyr spin-up with inversion, followed by 5 kyr without inversion. The spin-up
is fairly robust. There is some localized thickening and thinning (e.g., some thickening
and GL advance in Queen Maud Land around the Lazarev shelf), but overall, not much change
in grounding lines in key regions such as the ASE, Filchner-Ronne, and Ross.

I also fixed a minor error in the verbose diagnostics for Picard acceleration.

This commit is answer-changing only for runs with Laplacian smoothing during inversion.