Option to invert for coulomb_c at basin scale
Until now, inversion for coulomb_c has been local; the value at each vertex
is adjusted based on the thickness at that vertex.

This commit introduces an alternative, which_ho_coulomb_c_basin:
* The default is 0, implying no basin-scale inversion of coulomb_c.
* When which_ho_coulomb_c_basin = 1, CISM inverts for coulomb_c at the basin scale
  based on the basin-mean thickness of grounded ice.
* When which_ho_coulomb_c_basin = 2, CISM reads the basin-scale values from
  an external file (typically, these would be values from a previous inversion).

The logic is similar to that for the which_ho_deltaT_basin option,
but with tendency terms based on those for local coulomb_c inversion.

The thickness targets are given by the initial thickness of grounded ice,
interpolated to vertices. They are contained in a new field, grounded_thck_target,
that can be written to and read from restart files.

For summing and averaging vertex quantities over basins, I adopted the convention
that vertex (i,j) belongs to basin nb if and only if cell (i,j) belongs to basin nb.

In short test runs, the new inversion option seems to be working as desired.
I verified exact restart.