Moved subroutine point_diag to glimmer_utils
This commit moves subroutine point_diag from the glide_diagnostics module
to the glimmer_utils module to avoid circular 'use' statements.

The issue is that glide_diagnostics is not very low-level; it uses some glissade modules
from which it could be convenient to call point_diag.
In that case, point_diag needs to live elsewhere.